+++ src/constants.js
@@ -13,6 +13,7 @@ export const itemStatuses = {
   CLAIMED_RETURNED: 'Claimed returned',
   DECLARED_LOST: 'Declared lost',
   AGED_TO_LOST: 'Aged to lost',
+  LOST_AND_PAID: 'Lost and paid',
 };
 
 export const loanStatuses = {
@@ -21,6 +22,9 @@ export const loanStatuses = {
 
 export const loanActions = {
   CLAIMED_RETURNED: 'claimedReturned',
+  DECLARED_LOST:'declaredLost',
+  AGED_TO_LOST:'itemAgedToLost',
+  CLOSED_LOAN:'closedLoan',
 };
 
 // The names of the mutators which can executed on a given loan.
@@ -54,3 +58,38 @@ export const statusFilter = [
  with hardcoded ids for now. */
 export const feeFineBalanceId = 'cf7a0d5f-a327-4ca1-aa9e-dc55ec006b8a';
 export const MAX_RECORDS = '10000';
+
+export const paymentStatusesAllowedToRefund = [
+  'Paid fully',
+  'Paid partially',
+  'Transferred fully',
+  'Transferred partially',
+];
+
+export const reportColumns = [
+  'borrower.name',
+  'borrower.barcode',
+  'borrowerId',
+  'dueDate',
+  'loanDate',
+  'loanPolicy.name',
+  'loanPolicyId',
+  'loanId',
+  'feeFine',
+  'item.title',
+  'item.materialType.name',
+  'item.status.name',
+  'item.barcode',
+  'item.callNumberComponents.prefix',
+  'item.callNumberComponents.callNumber',
+  'item.callNumberComponents.suffix',
+  'item.volume',
+  'item.enumeration',
+  'item.chronology',
+  'item.copyNumber',
+  'item.contributors',
+  'item.location.name',
+  'item.instanceId',
+  'item.holdingsRecordId',
+  'itemId'
+];