+++ src/settings/PatronGroupsSettings.js
@@ -7,8 +7,6 @@ import {
 import { ControlledVocab } from '@folio/stripes/smart-components';
 import { stripesConnect, withStripes } from '@folio/stripes/core';
 
-import PatronGroupNumberOfUsers from '../components/PatronGroupNumberOfUsers';
-
 class PatronGroupsSettings extends React.Component {
   // adding the desired-count parameter, :50, to this query is an egregious
   // hack that willfully and knowingly abuses facets to contort them to
@@ -40,16 +38,7 @@ class PatronGroupsSettings extends React.Component {
   }
 
   render() {
-    const {
-      resources,
-      intl,
-    } = this.props;
-    const formatter = {
-      numberOfObjects: item => (<PatronGroupNumberOfUsers
-        item={item}
-        usersPerGroup={resources ? resources.usersPerGroup : null}
-      />),
-    };
+    const { intl } = this.props;
 
     return (
       <this.connectedControlledVocab
@@ -63,11 +52,11 @@ class PatronGroupsSettings extends React.Component {
         labelSingular={intl.formatMessage({ id: 'ui-users.information.patronGroup' })}
         objectLabel={<FormattedMessage id="ui-users.information.patronGroup.users" />}
         visibleFields={['group', 'desc']}
+        hiddenFields={['numberOfObjects']}
         columnMapping={{
           group: intl.formatMessage({ id: 'ui-users.information.patronGroup' }),
           desc: intl.formatMessage({ id: 'ui-users.description' }),
         }}
-        formatter={formatter}
         nameKey="group"
         id="patrongroups"
         sortby="group"