+++ src/views/UserEdit/UserEdit.js
@@ -8,14 +8,13 @@ import {
   cloneDeep,
   find,
   omit,
-  differenceBy,
   get,
   compact,
 } from 'lodash';
 
 import { LoadingView } from '@folio/stripes/components';
 
-import { eachPromise, getRecordObject } from '../../components/util';
+import { getRecordObject } from '../../components/util';
 
 import UserForm from './UserForm';
 import { toUserAddresses, getFormAddressList } from '../../components/data/converters/address';
@@ -231,7 +230,7 @@ class UserEdit extends React.Component {
     }
 
     if (permissions) {
-      this.updatePermissions(permissions);
+      this.updatePermissions(user.id, permissions);
     }
 
     if (servicePoints && stripes.hasPerm('inventory-storage.service-points-users.item.post,inventory-storage.service-points-users.item.put')) {
@@ -264,14 +263,42 @@ class UserEdit extends React.Component {
     });
   }
 
-  async updatePermissions(perms) {
-    const mutator = this.props.mutator.permissions;
-    const prevPerms = (this.props.resources.permissions || {}).records || [];
-    const removedPerms = differenceBy(prevPerms, perms, 'id');
-    const addedPerms = differenceBy(perms, prevPerms, 'id');
+  async updatePermissions(userId, perms) {
+    const {
+      permissions: permissionsMutator,
+      perms: permUserMutator,
+      permUserId,
+    } = this.props.mutator;
 
-    await eachPromise(removedPerms, mutator.DELETE);
-    await eachPromise(addedPerms, mutator.POST);
+    const { perms: permUserRecords } = this.props.resources;
+    // the perms parameter is an array of permission objects, but the permissions API
+    // wants an array of permission names.
+    const permissionNames = Object.values(perms).map(p => p.permissionName);
+
+    // If the user record has never had any associated permissions, a user permissions
+    // record may not exist. The PUT operation will fail if that's the case; thus,
+    // if no record is found, one has to be created before we assign the permissions
+    // as the last step.
+    if (permUserRecords.records.length === 1) {
+      const record = permUserRecords.records[0];
+
+      // N.B. permUserId is the id of the *permissions user* record, not the regular
+      // user record!
+      permUserId.replace(record.id);
+      record.permissions = permissionNames;
+
+      // Currently there is a bug (https://issues.folio.org/browse/MODPERMS-100) that
+      // requires us to remove the metadata object before sending the request.
+      delete record.metadata;
+
+      permissionsMutator.PUT(record);
+    } else {
+      // Create a new permissions user record first
+      await permUserMutator.POST({ userId }).then(record => {
+        record.permissions = permissionNames;
+        permissionsMutator.PUT(record);
+      });
+    }
   }
 
   render() {